Quan do 2

+++ Server/src/controllers/categoryController/category.js
@@ -2,9 +2,10 @@ const mongoose = require('mongoose');
 const CategoryModel = require('../../models/categoryModel/category');
 const ProductModel = require('../../models/productModel/product');
 const ImageProduct = require('../../models/imageModels/image');
+const SubCategory = require('../../models/subCategoryModel/subCategory');
 const categoryController = {
   addCategory: async (req, res) => {
-    const { name, description, position } = req.body;
+    const { name, description, order } = req.body;
 
     if (!name || !description) {
       return res.status(400).json({ message: 'Cần phải nhập tên và mô tả.' });
@@ -14,7 +15,7 @@ const categoryController = {
       const category = new CategoryModel({
         name,
         description,
-        position
+        order
       });
       await category.save();
 
@@ -31,6 +32,13 @@ const categoryController = {
       return res.status(400).json({ message: 'Cần phải có ID để xóa danh mục.' });
     }
 
+    const subCategories = await SubCategory.find({ category_id: id });
+
+    if (subCategories.length > 0) {
+      return res.status(403).json({
+        message: 'Không thể xóa danh mục vì nó chứa các danh mục con.'
+      });
+    }
     try {
       const deletedCategory = await CategoryModel.findByIdAndDelete(id);
 
@@ -92,14 +100,14 @@ const categoryController = {
 
   updateCategory: async (req, res) => {
     const { id } = req.params;
-    const { name, description, position } = req.body;
+    const { name, description, order } = req.body;
 
-    // if (!name || !description) {
-    //   return res.status(400).json({ message: 'Cần phải nhập tên và mô tả.' });
-    // }
+    if (!name || !description) {
+      return res.status(400).json({ message: 'Cần phải nhập tên và mô tả.' });
+    }
 
     try {
-      const updatedCategory = await CategoryModel.findByIdAndUpdate(id, { name, description, position }, { new: true });
+      const updatedCategory = await CategoryModel.findByIdAndUpdate(id, { name, description, order }, { new: true });
 
       if (!updatedCategory) {
         return res.status(404).json({ message: 'Không tìm thấy danh mục.' });
@@ -112,57 +120,58 @@ const categoryController = {
     }
   },
 
-   getProductWithCategory: async (req, res) => {
+  getProductWithCategory: async (req, res) => {
     const { id } = req.params;
-    if (!id) return res.status(404).json({ message: "Không có id" });
+    if (!id) return res.status(404).json({ message: 'Không có id' });
 
     try {
-        const categories = await CategoryModel.aggregate([
-            {
-                $match: { _id: new mongoose.Types.ObjectId(id) }
-            },
-            {
-                $lookup: {
-                    from: 'subcategories',
-                    localField: '_id',
-                    foreignField: 'category_id',
-                    as: 'subcategories'
-                }
-            },
-        ]);
-
-        if (!categories || categories.length === 0) {
-            return res.status(400).json({ message: "Không tìm thấy danh mục con" });
+      const categories = await CategoryModel.aggregate([
+        {
+          $match: { _id: new mongoose.Types.ObjectId(id) }
+        },
+        {
+          $lookup: {
+            from: 'subcategories',
+            localField: '_id',
+            foreignField: 'category_id',
+            as: 'subcategories'
+          }
         }
+      ]);
 
-        const subcategoryIds = categories[0].subcategories.map(subcategory => subcategory._id);
+      if (!categories || categories.length === 0) {
+        return res.status(400).json({ message: 'Không tìm thấy danh mục con' });
+      }
 
-        const products = await ProductModel.find({ sub_category_id: { $in: subcategoryIds } });
+      const subcategoryIds = categories[0].subcategories.map((subcategory) => subcategory._id);
 
-        const productsWithImages = await Promise.all(products.map(async (product) => {
-            const images = await ImageProduct.find({ product_id: product._id });
-            return {
-                ...product.toObject(), 
-                images: images 
-            };
-        }));
+      const products = await ProductModel.find({ sub_category_id: { $in: subcategoryIds } });
 
-        const filteredCategories = categories[0].subcategories.map(subcategory => {
-            const productsForSubcategory = productsWithImages.filter(product => 
-                product.sub_category_id && product.sub_category_id.toString() === subcategory._id.toString()
-            );
-            return {
-                subcategoryId: subcategory._id,
-                products: productsForSubcategory
-            };
-        });
+      const productsWithImages = await Promise.all(
+        products.map(async (product) => {
+          const images = await ImageProduct.find({ product_id: product._id });
+          return {
+            ...product.toObject(),
+            images: images
+          };
+        })
+      );
 
-        res.status(200).json(filteredCategories);
+      const filteredCategories = categories[0].subcategories.map((subcategory) => {
+        const productsForSubcategory = productsWithImages.filter(
+          (product) => product.sub_category_id && product.sub_category_id.toString() === subcategory._id.toString()
+        );
+        return {
+          subcategoryId: subcategory._id,
+          products: productsForSubcategory
+        };
+      });
 
+      res.status(200).json(filteredCategories);
     } catch (error) {
-        res.status(500).json({ message: error.message });
+      res.status(500).json({ message: error.message });
     }
-}
+  }
 };
 
 module.exports = categoryController;
